requirements.txt
requests==2.22.0
google-cloud-dns==0.31.0
